+++ release_notes.md
@@ -19,9 +19,17 @@
 
 ### Enhancements:
 
-* Added support for encryption.
-* Adding SharedGroup::wait_for_change() which allows a thread to sleep until the
-  database changes.
+* Added SharedGroup::get_number_of_versions() which will report the number of
+  distinct versions kept in the database.
+* Added support for encryption
+* Adding SharedGroup::wait_for_change() which allows a thread to sleep until
+  the database changes.
+
+-----------
+
+### Internals:
+
+* Lorem ipsum.
 
 ----------------------------------------------